refactor(components/atom/actionButton): 1 single component per file
