+++ components/atom/actionButton/README.md
@@ -1,6 +1,6 @@
 # AtomActionButton
 
-SUI Atom ActionButton
+SUI Atom ActionButtonWrapper
 
 ![Atom Action Button Colors](./assets/action-button-colors.png)
 ![Atom Action Button Sizes and States](./assets/action-button-sizes-states.png)
@@ -16,7 +16,7 @@ $ npm install @s-ui/react-atom-action-button
 ### Basic usage
 
 ```js
-import ActionButton from '@s-ui/react-action-atom-button'
+import ActionButtonWrapper from '@s-ui/react-action-atom-button'
 
 const Icon = () => {
   return (
@@ -25,9 +25,9 @@ const Icon = () => {
 }
 
 return (<div>
-  <ActionButton icon={Icon}>Normal</ActionButton>
-  <ActionButton icon={Icon} focused>Focused</ActionButton>
-  <ActionButton icon={Icon} size='large' disabled>Disabled, Large</ActionButton>
+  <ActionButtonWrapper icon={Icon}>Normal</ActionButtonWrapper>
+  <ActionButtonWrapper icon={Icon} focused>Focused</ActionButtonWrapper>
+  <ActionButtonWrapper icon={Icon} size='large' disabled>Disabled, Large</ActionButtonWrapper>
 </div>)
 
 ```
